# server-go/docs/swagger.yaml
definitions:
...
  domain.AdminFilterTournament:
    properties:
      clubId:
        type: string
      clubName:
        type: string
      id:
        type: string
      name:
        type: string
      organizatorFirstName:
        type: string
      organizatorId:
        type: string
      organizatorTelegramId:
        description: Дополнительные поля для удобства фильтрации
        type: integer
      tournamentType:
        type: string
    type: object
...
  domain.AdminPatchTournament:
    properties:
      clubId:
        type: string
      courtId:
        type: string
      description:
        type: string
      endTime:
        type: string
      maxUsers:
        type: integer
      name:
        type: string
      organizatorId:
        type: string
      price:
        type: integer
      rankMax:
        type: number
      rankMin:
        type: number
      startTime:
        type: string
      tournamentType:
        type: string
    type: object
...
  domain.Court:
    properties:
      address:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  domain.CreateTournament:
    properties:
      clubId:
        type: string
      courtId:
        type: string
      description:
        type: string
      endTime:
        type: string
      maxUsers:
        type: integer
      name:
        type: string
      organizatorId:
        type: string
      price:
        type: integer
      rankMax:
        minimum: 0
        type: number
      rankMin:
        minimum: 0
        type: number
      startTime:
        type: string
      tournamentType:
        type: string
    required:
    - clubId
    - courtId
    - maxUsers
    - name
    - organizatorId
    - startTime
    - tournamentType
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.Loyalty:
    properties:
      description:
        type: string
      discount:
        type: integer
      id:
        type: integer
      name:
        type: string
      requirements:
        type: string
    type: object
  domain.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.PlayingPosition:
    enum:
    - right
    - left
    - both
    type: string
    x-enum-varnames:
    - PlayingPositionRight
    - PlayingPositionLeft
    - PlayingPositionBoth
  domain.Registration:
    properties:
      date:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/domain.RegistrationStatus'
      tournamentId:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      userId:
        type: string
    type: object
  domain.RegistrationStatus:
    enum:
    - PENDING
    - ACTIVE
    - CANCELED
    - CANCELED_BY_USER
    type: string
    x-enum-varnames:
    - RegistrationStatusPending
    - RegistrationStatusActive
    - RegistrationStatusCanceled
    - RegistrationStatusCanceledByUser
...
  domain.Tournament:
    properties:
      clubId:
        type: string
      court:
        $ref: '#/definitions/domain.Court'
      description:
        type: string
      endTime:
        type: string
      id:
        type: string
      maxUsers:
        type: integer
      name:
        type: string
      organizator:
        $ref: '#/definitions/domain.User'
      participants:
        items:
          $ref: '#/definitions/domain.Registration'
        type: array
      price:
        type: integer
      rankMax:
        type: number
      rankMin:
        type: number
      startTime:
        type: string
      tournamentType:
        type: string
    type: object
...
  domain.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      birthDate:
        type: string
      city:
        type: string
      firstName:
        type: string
      id:
        type: string
      isRegistered:
        type: boolean
      lastName:
        type: string
      loyalty:
        $ref: '#/definitions/domain.Loyalty'
      padelProfiles:
        type: string
      playingPosition:
        $ref: '#/definitions/domain.PlayingPosition'
      rank:
        type: number
      telegramId:
        type: integer
      telegramUsername:
        type: string
    type: object
...
info:
  contact: {}
  description: Manage chats, users, tournaments. For JWT authentication use 'Bearer
    <token>' format in Authorization header.
  title: GoPadel server
  version: "1.0"
paths:
...
  /admin/tournaments:
    post:
      consumes:
      - application/json
      description: Create a new tournament. Available for any admin.
      parameters:
      - description: Tournament data
        in: body
        name: tournament
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTournament'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Tournament'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create tournament (Admin)
      tags:
      - admin-tournaments
  /admin/tournaments/{id}:
    delete:
      description: Delete tournament. Available for any admin.
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete tournament (Admin)
      tags:
      - admin-tournaments
    patch:
      consumes:
      - application/json
      description: Update tournament data. Available for any admin.
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Tournament update data
        in: body
        name: tournament
        required: true
        schema:
          $ref: '#/definitions/domain.AdminPatchTournament'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tournament'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update tournament (Admin)
      tags:
      - admin-tournaments
  /admin/tournaments/filter:
    post:
      consumes:
      - application/json
      description: Get all tournaments with advanced filtering. Available for any
        admin.
      parameters:
      - description: Filter criteria
        in: body
        name: filter
        schema:
          $ref: '#/definitions/domain.AdminFilterTournament'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Tournament'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tournaments (Admin)
      tags:
      - admin-tournaments
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Token
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"